requirements.txt
pytorch
sklearn
numpy
tqdm
pygame